AST Transformation for Remote DBus Adapters
Currently used by sbml and user, not yet for model (needs adjustments).

The AST Transformation changes code in a plugin and that does not work as the code for
the transformation needs to be compiled before other code and all plugins are compiled at
the same time.

Because of that a new ant task is added: ast-jar
This compiles the ast sources, generates a jar file and puts it into the core's and
web application's lib folder. This ant task has to be run once whenever the ast source
code changes.

I will try to integrate it into our grails-core script and make the other ant tasks
depend on it, so that the user has not to care about it.