Simplify retrieval of lists of elements through DBus
With DBus we can only retrieve a list of identifiers and we need to manually fetch the
further elements. This commit adds a method to fetch further data in a generic way using
groovy dynamic features. The idea is to generate the code to invoke the method in the
AST transformation in the next step.