AST transformation for DBusAdapter
The AST transformation for the DBusAdapter differs slightly from the other transformations.
The methods for which the code has to be generated by the transformation, have to be implemented
and annotated with the DBusMethod annotation. This annotation is required to mark whether a
method needs to set the authentication and will be extended to map the method names to the
name of the service.

It would be nice to have this annotation in the interface, but this seems not to be possible.

In the first step some methods of the ModelDBusAdapter are replaced by the transformation.