Resolve all annotations when a Model is uploaded
The FileFormatService is extended to provide a method to retrieve all URNs used in a model file
which is implemented for SBML in the SbmlService. The MiriamService is adjusted to retrieve all
names for such a list of URNs. Therefore the service is adjusted to no longer try to retrieve
the name when just accessing the stored data.

The new way to retrieve the names is invoked from ModelService when a new model revision is
uploaded or a new model is uploaded. This should be moved into a thread (as it involves an
initial parsing of the Model file).