Fix loading of of annotations after a Model is uploaded
The loading of annotations is done in a thread which leads to a problem in relation to
Hibernate session. The model loading thread creates the Annotation loading thread
before the transaction is committed to the database (which happens after the method
returns). The Annotation loading thread uses an own Hibernate session (which is important
so that Exceptions won't roll back the loaded Model). Due to the concurrent situation it
is likely that the Annotation thread is scheduled before the Model loading thread has
committed the transaction. In that case accessing the Model from the Annotation thread
is just not possible.

The solution to the problem is at least in this case quite simple: if we could not get
the Model in the Annotation thread we let the Thread sleep some seconds and try again.
This makes it likely that the transaction is committed and we can access the Model.