Store relationships between Gene Ontology terms
For each resolved MiriamIdentifier of the type Gene Ontology a further
domain object GeneOntology is created and the relationship to other
Gene Ontologies is resolved. Therefore the GeneOntologyResolver is
used and changed to cache the already parsed XML.

To ensure that the right XML is used the GeneOntologyResolver is of
scope prototype so that each resolve operation gets it's own resolver.