Properly synchronize the storing of MiriamIdentifiers
Solves the issues with concurrent access to the MiriamIdentifier table.
The FetchAnnotationsThread now only fetches the URNs from the Model files
and passes them to the MiriamService for scheduling. The MiriamService
uses a lock to ensure that only one thread will resolve the name and
schedules a new ResolveMiriamIdentifierThread for a URN which has not
yet been scheduled.

The MiriamIdentifiers are now properly resolved without any thrown
hibernate exception.

The change does not yet incorporate the GeneOntology resolving. This
is at the moment broken.