Create relationships between gene ontologies
Whenever the GeneOntologyResolver finds a relationship it schedules
this with the from, to and type information in the MiriamService.
The MiriamService temporarily stores the information in a List and
schedules the to gene ontology for Miriam name resolving.

As soon as a gene ontology is discovered it will be tried to create
all relationships with this gene ontology. A loop through the list
is performed and if to and from gene ontologies exist, the relationship
is created.

The procedure seems to work correctly as the relationships are generated
in a way, so that there are only three gene ontologies without a furhter
gene ontology.