Publish a Model Revision
Allows a user with role ROLE_CURATOR and write permission on a Revision to
publish it, which adds read permission to ROLE_USER and ROLE_ANONYMOUS.

Only core part is implemented, remote service not yet implemented as curation
currently missing. Main reason for adding was to make it available from the
DBus interface.

Needs a unit test. Currently only administration access is tested.