Switch to DBus System Bus
The DBus Session bus is linked to a "user" session, which implies an X11 Display.
This does not make sense for a server application like JUMMP.

Because of that JUMMP now uses the System DBus. In order to allow JUMMP to register
to the system DBus the example config file net.biomodels.jummp.conf has to be copied
to: /etc/dbus-1/system.d/

The first policy contains the name of the user which is allowed to own the bus. This
might be adjusted to the local situation.