Support both DBus SESSION and SYSTEM bus
A new boolean config option "jummp.plugins.dbus.systemBus" is added which decides
whether the core exports to the system bus (true) or to the session bus (false).
By default the session bus is currently set.

In the remote part peer remote objects are resolved to ensure that an exception
is thrown if the object is not available on a given bus. First it is tried to
resolve the object on the System bus. If this fails the session bus is querried.
The config option is not used as this would result in a chicken egg problem.
In order to retreive the config option DBus is required to query the core application
for the bus option.