Add required jdom and jaxen dependencies
Since Grails 2 these libraries seem to not be included any more.
There seems to be no maven repository for jdom version 1.1.1 and
including jaxen 1.1.1 through maven destroys the generation of
unit test results.

Have I mentioned that the dependency handling of Java is far from
optimal to put it nicely?