Fix testRetrieveModelFile integration test
The test ensured that no file can be retrieved where the revision has not been
created through the ModelService. This test started to fail since the retrieveFile
method in VcsService does no longer interact with the VCS in case the revision is
the "HEAD" revision which was the case in such a situation.

The test is adjusted to import a further revision to make the manually created not
the latest.