Handle Jummp Plugin dependencies through jar files
Since grails 2 it is possible to package a plugin as a jar:
grails package-plugin --binary

This allows a better dependency handling. A new ant task "packagePlugins"
is added to generate all jar files and copy them to a shared directory
"pluginlibs" in the jummp base directory. This directory is used as a
repository for jars in all plugins and core and web-app.

Also all jars in lib subdirectories of plugins are moved into this
shared directory to better handle it.

In order to generate the jars either run the ant task or run the
package command in each plugin and manually move the jar to named dir.
It is important to know that grails will resolve the jar at runtime
and keep it in a chache. As we do not change the version number of the
plugins a change in a plugin will not be picked up. The existing files
have to be manually removed from the cache, which is located in:
~/.ivy2/cache

I will investigate how to get this updated in a more convenient way.

At the moment this is our "Break JUMMP" commit. Several functionality
is no longer working, including:
* JMS
* Several DBus services (GeneOntology still working)
* Probably more

This will be fixed with the next commits!