JmsAdapterService needs to delete models uploaded to the core
Uploading a Model through JMS is executed in another thread with its
own hibernate session. This means the transaction of the test will not
roll back that change. Because of that we have to delete all models again
through JMS to get the deletion also in an own session.

fixes #2