Ensure that bives diff dir is set correctly
If the config option is not set a default directory depending on the
Servlet Context should be used. Unfortunately during the configuration
the servletContext is not yet available.

To circumvent this problem the option is read again in diffDataService
and changed if needed. All bives classes which access the config option
now use the diffDataService to retrieve the directory which ensures
that the diffDataService is always created before the config option
is used.