Implement a Model History functionality
Whenever a user accesses a Model a history item is created through
the ModelHistoryService. This allows to retrieve a list of the last
accessed models by a user. The size of the list is configurable with
0 meaning feature is completely disabled and nothing is stored.

Currently there is nothing in the web frontend yet to see the list,
the complete functionality is ensured through the unit test.
With Grails 2 a new way of testing is introduced which required to
duplicate some code which is also in JummpIntegrationTestCase.
Solution: port the existing integration tests to the new way.