Adjust AST for newer Groovy version
With newer Groovy version the transformation has problems when just re-using
a ClassNode from the interface in case it uses Generics.

To make that work again the transformation generates a new ClassNode of same
type unless it is a plain old data type.

For enums as parameters the transformation is also failing, as a workaround
the methods using such enums are no longer generated.