Initial import of new layout
The layout is implemented in the Sitemesh layout and a less file.
For less the grails plugin is used. The style is defined in the core
application as it is difficult to get this right for the complete
application if it is in a plugin.

It is not implemented in the web application as the idea is to use
a plugin for the user interface which could then either be used
directly in core or keeping the two application approach.

Using a plugin for the web frontend makes it easy to replace the
complete user interface by a different implementation.