Integrate Weceem Content Management plugin
Why does JUMMP need a CMS? JUMMP has quite some content areas which are
specific to the installation. Think of imprint, branding, funding bodies,
news, etc. etc.

Writing custom code for all these things is quite some work especially if
there is a CMS plugin for Grails. Using such a plugin seems to be more
straightforward.

The Weceem plugin is integrated in a way that e.g. the imprint is read from
a page managed in the CMS. Only an admin user has access to modify anything
in the CMS.

A more fine grained user management system can be integrated by modifying
scripts/weceem-security.groovy. This file is sourced at startup. Whether this
works with war has to be tested.

Unfortunately the binding between spring security and weceem is rather
broken. In fact the weceem plugin is completely unable to handle users
authenticated through LDAP. Because of that there is some glue code in
Bootstrap.groovy. Even with that it can happen that actions run on error,
but they are performed nevertheless.

It should be possible to define some default templates which get installed.
This has to be evaluated.