Adaptions for database, branding and new access method of config
The configuration properties have to be accessed via
grailsApplication in Grails 2.0 and the old method
does not work anymore.
As JUMMP now supports PostgreSQL in addition to MySQL
the MysqlCommand is replaced by DatabaseCommand
having an additional field for the database server's
type.
To support changing colors of i.e. fonts in the
web front-end support for this feature is added
through the BrandingCommand.