Fix i18n issues with the jquery plugin
To work correctly the messages.properties need to be in the javascript
folder. This seems to only work for the core application but not in a
plugin. Due to that some messages are moved to the core and during
compile the current messages.properties is copied into the web-app
folder.