Improve switching between Model tabs
Link to the data which is loaded is placed in the tab's div as
a rel attribute. When the overlay is shown click handler on the
divs are registered. This is better than using a link as the
complete tab becomes clickable.

The tab content is loaded based on the rel element and the classes
of all tabs are adjusted to have a "selected" class only on the tab
on which the data could successfully be loaded.

On show of overlay the first tab is loaded.