Filter GeneOntology tree based on search terms
An autocomplete list is added to the GeneOntology tree. This list
auto completes gene ontologies based on name or id. If a suggested
element is selected the GO-Tree starts to load all nodes to the
selected ontology and expands the tree, so that the Gene Ontology
node is selected and expanded.