Adding the world's most simple search
Using datatables search functionality to have a google like
search field.

The actual search is performed on the server by appending the
HQL query to match name of model or journal, title or
affilitation of the publication.

The search is only performed if there are at least three
characters to not return everything. Results are of course
paginated just like any other results for the datatable.

Possible extension: search on authors.