Disabled JMS and Quartz to prevent memory leaks when JUMMP is stopped in Tomcat.
Commented out the AuthenticationOutdaterJob job as it relied on Quartz. We should
re-implement the functionality differently. Our previous Weceem security policy,
stored in scripts/weceem-security.groovy, threw ClassFormatErrors. For now, use
the default wcm security policy.

Weceem lists Quartz 0.4.2 as a dependency and uses it internally. Be sure to edit
~/.grails/$version/jummp/plugins/weceem*/ and comment out the relevant sections.