Explicitly state from where the git Repository should find the git dir
Since JUMMP migrated to git the command findGitDir returned the wrong
git directory if grails is executed from the working tree. The reason
is that findGitDir starts to look for the git directory starting from
the directory where the JVM is executed. This of course finds the git
directory of the local clone of JUMMP.

The result is that neither the unit tests nor run-app can work with
the git repository resulting in models not being able to import and all
tests failing.

This change fixes this issue by always specifying from where findGitDir
should start looking for the git directory.