Configured default package settings.JummpPluginConfig location is now consistent.
Now when an artifact gets created, it receives the correct fully qualified name
For example, a new controller in the web application plugin will be part of the
net.biomodels.jummp.webapp package, a new service in core will be created in
net.biomodels.jummp.core.

Also ensured that JummpPluginConfig lies in the same package as the other Groovy
classes within a given plugin.