Made all JUMMP plugins inline. Corrected the loading mechanism of our plugins.
Any configurations needed by a plugin are performed in its plugin descriptor's
doWithSpring clojure instead of in a separate class. Also added constraints to
the loading order of the plugins so that dependencies are met.

This reduces dependency resolution times, which means that most grails commands
execute quicker because they no longer involve checking for the latest version
of the plugin JARs. It also boosts productivity, as changes to any of the plugins
are now picked up by the application without the need to restart it with refreshed
dependencies. As a bonus, this also enables the use of grails in interactive mode.

AST.jar is now added to the classpath from the DBus plugin rather than the main app,
which should have a positive impact on the execution time of grails commands. However,
this means that when we create the jar, we need to ensure that we use the correct relative
basepath to jummp-plugins/jummp-plugin-ast/src/groovy, otherwise you risk not being able
to compile the DBus plugin as a standalone application.

Finally, all our JAR dependencies have been moved to the lib/ folder for
added efficiency. Since the pluginlibs folder is now redundant, it has been
removed.