Made JummpPluginGit run properly. Potential ServletContext issue.
It needs grails-app/views/index.gsp to avoid a 404.

Added the logic to parse ~/.jummp.properties in Config.groovy,
as they cannot be loaded from JUMMP's settings. It seems that we
cannot inject the servletContext bean unless we wire it ourselves
in the init closure of BootStrap, so the plugin descriptor should
use a MockServletContext.