Amended dependencies, bean definitions and logging for jummp-plugin-subversion.
Introduced missing SVNKit dependency in BuildConfig. The bean definitions from
the plugin descriptor's doWithSpring closure are now copied into resources.groovy
so that the plugin can be run as a standalone application. The new logs/ folder now
contains the logging for the subversion plugin, again useful for when we run it as a
Grails application, rather than a JUMMP plugin.