Added a FileSystemService to compartmentalise the folder where models are stored.
This service is responsible for providing a flexible and scalable layout for
the folder where models will be stored. It splits the working directory into
containers that can hold a configurable number of models. This facilitates the
clustering of the files that belong to a model, e.g. SBML, SedML, SBGN, pdf.

It also mitigates against the risk of increased file access time due to the
working directory containing over 2-3000 files, which is the upper limit of
most file systems.

The naming convention of the containers, defined by the constant
containerPatternSeed, is currently set to be three characters long e.g. aaa,
aab,...,zzz and, with a current threshold of 1000 model folders per container,
we get a maximum of 26^3 * 1000 = 17576000 models per instance, each holding an
arbitrary number of files.