Introduced a domain object responsible for storing some metadata about files.
These files belong to one or many Versions of a given Model. The domain object
records the location of the file relative to the root folder where all models
are deposited, a brief description of the relation file to the model as well as
whether the file should be displayed to the user or not in the web interface.