Fixed SBMLService, ModelDelegateService and IModelService retrieveModel functions
Updated code, and passed two integration tests that were failing because of SBML service. One apparent concern is that groovy closures were not being executed, possibly because the lists/maps in question are imported as java.util.x . This was done because certain interfaces are defined in Java classes. If this is correct, then SBMLService and other services should therefore use conventional iterators over the lists/maps instead of each {} and find {} etc.