ModelService Integration tests passed
Integration tests passed for model service. However they required
disabling of Bives revision created listener as it was retrieving
the same files from VCS as were being tested. Current implementation
does not use a different name for each retrieve call resulting
in tests failing for when the same named file of different revisions
was retrieved by clients of the vcs API. The simple fix is to create a different
temp directory within the exchange dir for each vcs retrieve call.
This will require rewriting of the tests.