Tests passed, SbmlService tests and Bives issues fixed
GitManager writes to a temporary directory (within exchange)
for each retrieve call. Revision events are generated with
retrieved files from revisions instead of passing the original
files around (which were being read by Revision Created Listener
resulting in on-and-off SbmlService test failure).

Revision Created Listener has the responsibility of deleting the
temp files created for Revision Created Event. If there is any
other listener for this event, we need to reconsider how to
delete the files!