SbmlService tests modified to have one model file per directory
Current implementation of SbmlService returns the first file
in the model directory that is a valid Sbml document. This is
problematic for cases where multiple model files are in the
directory. We need a non-arbitrary way of getting to the model
file we want, instead of getting the first one (or most recent, etc).
For now the tests are written to have one model file per directory.