Test for Concurrent Access to Model storage added
New test checks whether two threads beginning at the same time
can access VcsManager concurrently if they are accessing different
models. This is done by having one read a small model, and another
write a large one, and asserting whether the former terminates before
the latter. The reliability of the test is increased as the file size
for the second thread is increased. However, the execution time of the
test is also correspondingly increased.