Changed RepositoryFile wrappers to always expect absolute paths.
The domain object is in charge of capturing the correct subpath of it and
store that in the database. This process is symmetrical, that is, the domain
object also turns the relative path from the database into an absolute one when
an instance of this domain object is wrapped in a transport command.