Change to RepositoryTransportCommand creation mechanism
When RepositoryFiles are retrieved from the database, their
path may be referring to what it was during insertion (i.e.
the file in the model directory). However when transport commands
are created from these objects, the path cannot be that, as it
is assumed to correspond to the file belonging to that particular
revision (instead of as it currently is in the model directory).
To get around this problem, the 'toCommandObject()' function of
RepositoryFile has been disabled, and instead the Revision domain
object is responsible for retrieving the correct files from the
VcsManager, and associating the paths of these files with the
RepositoryFileTransportCommand objects it creates