Allow submission to proceed with empty submission if main file exists
Restructured the submission flow to allow submissions when
the user hasnt uploaded any files, but there are valid files
associated with the model. This can happen when the user either
goes back to the file upload state, and wants to move forward
without uploading a new file (they changed their mind) or during
the update model flow, when the user doesnt want to add a new
file, but wants to change the name/description etc.