Using file based locking for inter-process GitManager use
GitManager uses reentrant (memory based) locks for managing
access to model working directories, to prevent conflicts.
While this works across threads, it doesnt work across processes.
To deal with this, threads acquire file locks. This is in addition
to the reentrant lock, because file locks alone are not a wholly
predictable mechanism. Firstly they are only 'advisory', and implementations
vary across OSs. Secondly, accessingly the same lock within the same
process (which would be the case for multi-threaded access) results
in an exception instead of a block (even with tryLock). Therefore
both in-memory and file locks are used to manage access