The real multi-process testing commit. Last one was a merge.
New tests access and write to the same repository from
multiple threads and processes. Tests pass if each thread and
process always gets the file it wrote to the repository. As process
and thread scheduling is non-deterministic, its difficult to
ensure correctness. Therefore the simultaneous writing is
done 200 times across three threads and 1 separate process.
It is hoped that any errors will be uncovered.