Improved garbage collection settings.
Migrated to ConcurrentMarkAndSweep in conjunction with ParNewGC for
tenured and young objects respectively. Unlike the ParallelGC, CMS is
not a stop-the-world algorithm which results in increased responsiveness.
ParNewGC is also more sophisticated than ParallelGC, hence there is a
visible increase in the efficiency of the garbage collector, which means
that JUMMP can spend more time doing its job.

Future Improvements
    Investigate whether we can lower the maximum heap size to 2GB or so.
	VisualVM consistently shows PermGen/OldGen below 50%.
    Compare the current gc settings against Java 7's new G1 algorithm.