Corrected locking mechanism in GitManager.
Retrieving a list of revision IDs was implemented without
locking as it is used from within GitManager by other
functions that have already acquired the lock. As FileLocks
are not re-entrant, acquiring a lock again throws an
exception. However, this leaves accessing the list of revisions
from outside GitManager thread and process unsafe. Therefore
a private implementation with optional locking has been added.