First commit for the PharmML plugin.
Added a basic implementation of the FileFormatService interface. Several
models encoded in PharmML are available in the plugin's test/files folder. Note
that this plugin depends on jummp-plugin-sbml due to the fact that the structural
model may be encoded in SBML.

Also introduced a utility class in the core-api plugin that facilitates the retrieval
of elements from a given XML document.