Added comment for weird service injection mechanism in lazy loading mechanism
RevisionTransportCommands (RTC) now loads the file list only when someone
requests them. This saves on unnecessary and expensive retrievals from vcs when
iterating through a list of RTCs (as in the history tab in model display)
and not requiring the files.

Dependency injection doesnt work as usual in src/groovy. To explain the unusual
way of accessing modelDelegateService, the link to grails faq has been added.