Clean up resources from exchange after they are no longer in use
Added two quartz jobs for removing unused files from the exchange.
The first, based on the garbage collector's reference counting, polls
a reference queue (from the model delegate service), and removes
files associated with a revision TC marked by the GC. The second is a
backup in case the GC doesnt mark a revision for any reason (GC's garbage
collection policies, tomcat is restarted, the service is unloaded and so on).