Adding link verfiication to pubmedservice. Maybe cause to rename service?
Service used to be just for pubmed lookups. Now that we (aim to) support
identifiers.org sources and in time other services, perhaps it would be
good to redefine this service as one used for handling publications. Currently
publication related code is spread across multiple locations which isnt
ideal design.