Moving model state from model to revision
Moved model state from model to revision. As model state
was performing a dual function of recording the various states
of a model (appropriately placed in revision) and also recording
where the model as a whole was deleted (appropriate in model), an additional
parameter was added to model to record its deleted-ness. Tests updated
to reflect this change.

This change requires a drop of the database.