Adding the license and permission
License added to every file (groovy/java/gsp.. todo: javascript).
Also added generic permissions text to any file that had an
import that wasnt net.biomodels. Need to go through everywhere
where there is this text and see which license needs to be referred
and which plugin needs to be named, if any.