Slight refactoring of submission flow and better handling of exceptions
Broke the upload file state further, better separating file upload from
validation. Also added a general exception handling mechanism where
a 'ticket' is created, and the stack trace of the exception is
emailed to the configured admin address. Should handle any exception
generated in an 'action' state, although there still remain some
vulnerable parts of the code that are in the transition states. Dont
know how to throw exceptions from there (e.g. what happens if theres
a problem in the bindData call). Possibly something we need to live with.