Refactoring the model controller a bit more
Mainly to get the tests to stop sending emails. The redirect
doesnt work in the tests (works in real life) triggering an
exception which got mailed to admin (although the tests pass
and the model is updated). To avoid this pain, the redirect
 has been separated to a part of the code that doesnt send
mails.

Also more modular as a result.