Adding the user profile functionality
A basic profile page created with links to edit user details
and edit password pages. Added links in ddmore/biomodels layout.
Jstree.css contains quite a bit of css that should probably be included
in whatever css we decide to add to the project. Also all of the user
pages have a duplicated notification mechanism (in common with several
other parts of the code). This has been deliberately done, so that these
bits of code form the kernel of a generic notification system, a task for
which already exists.