Added generic notification mechanism
Replaced the various ways of giving error messages/notifications
with a single generic mechanism which is shown if there is either a string
(named flashMessage) or a bean with error message (named validationErrorOn).
Mechanism can also be used with client side calls (as in the user admin pages).
Added to ddmore layout so far, needs to be added to biomodels layout as well.