Fix to premature deleting of model files to be updated
GC based exchange cleaner deletes files from exchange based
on if references to the associated revision no longer exist.
It appears that references in flow memory confuse the GC,
resulting in the files being prematurely deleted during the
update flow. To rectify, a reference with a unique name is
stored in the session memory, and removed when the update
flow is terminated (either successfully or not)