Moved to Maven for dependency resolution.Also disconnected Bives plugin
Ivy is no longer the preferred option for dependency resolution. Maven
does not work with flatDir or SVN-based artifact repositories. The Bives
plugin had to be disconnected because it uses obsolete JARs which cannot
be pushed to a Maven repository.

Also upgraded to version 3.0 of the Servlet specification. This
preserves compatibility with Tomcat 7 while also allowing us to use the
async API.

Finally, I configured JUMMP and all plugins to run in forked mode in all
but the test environment as described in the Grails upgrade notes.