Adding support for web-service: /model/{id}?format=xml/json
The mechanism followed here is to implement classes reflective
of the structure we want in response to the service call. This saves
us from writing separate marshallers for JSON/XML. As the output
of the service uses a number of different transport command objects
and necessitates a modelService call, this makes more sense to me
currently.  On the other hand we arent using the marshalling/rendering
framework provided by grails, so it is a less standard approach. Please
consider this a prototype of the approach possible, and a perfectly
good candidate for further changes.