Refactored PharmML display generation in order to cope with multiple versions.
Introduced a dedicated interface for rendering PharmML representations
in the core-api plugin which is currently realised by the
PharmMl0_2AwareRenderer. A factory class exists for all implementations
of the above-mentioned interface.

The parts of PharmMlTagLib dealing with rendering the independent
variable, function defintions, structural- and variability models have
been moved to the PharmML 0.2 rendering service which populates a number
of GSP templates.

There is a dedicated service for managing the interaction between the
taglib and the various IPharmMlRenderer implementations, namely
PharmMlRenderingService, which is wired as a Spring bean in both the
pharmml plugin, and the main application, via the plugin descriptor.

PharmMlController now exposes the version of the PharmML representation
via the eponym variable in the model that is passed to the
format-specific view.

Also improved the documentation for IPharmMlService and removed an
outstanding to-do item in PharmMlVersionAwareHandlerFactory.