More flexible external file linking mechanism
This is a workaround for a limitation in how
external files are specified in the pharmml
language. We had earlier agreed that only one
markup would specify the file location, but currently
there is name, url and format. The url seems to be
for indicating the path to the file, and is therefore
ignored. The previous code was expecting the name to be the filename
but this doesnt work, as the name field cannot
contain a '.'. The intended behaviour seems to be to
use the name and the format and infer the full filename
(which is risky as CSV may not necessarily have a .csv extension).
This is likely to be resolved in the next pharmml
release. As a temporary workaround, a more 'intelligent'
search is performed through the additional files.