Support for PharmML 0.3.1.
Updated the version of libPharmML that we use, registered 0.3.1 as a new
format and linked it to the corresponding components responsible for
handling and rendering models encoded in this format.

Also updated the unit tests to work with 0.3.1 files instead of 0.3,
with the exception of those for ObservationEventsMap, as there are no
StudyEvent elements inside any of the 0.3.1 examples.