Added customisable model identifier generators.
This commit introduces a new package -
net.biomodels.jummp.core.model.identifier.generator - which comprises of
an interface and two implementations: one that always returns a null
model identifier, and another one which uses a collection of decorators
to build the id dynamically.

Atop of these classes sits a non-transactional Grails Service which is
in charge of registering and naintaining all model id generators
employed by the application at runtime.