Don't make ModelIdentifierGeneratorRegistryService a Grails Service.
We need to set the registry property, and being a Grails Service means
that the bean is wired by Grails automatically without setting it. To
get around this, one can either write a BeanFactoryPostProcessor which
unregisters Grails' definition of the bean and then registers it
properly, or just wires the bean manually.