Added a utility class for wiring up model id generators at runtime.
ModelIdentifierUtils methods are called very early on in the bootstrap
process, before the application context is set up. For this reason, this
class cannot rely on dependency injection, access to the database via
the dataSource bean or GORM.

The utility methods are geared towards defense, in the sense that they
throw exceptions and fail early in case of issues detected with the
model identifier generation settings. Having said that, one issue which
is not addressed is the change of the model id generation settings while
the application is shut down. Upon restart, we look in the database to
see what were the last identifiers that were generated, but cannot
accurately detect whether they match the patterns defined in the config.

Also improved the robustness of ModelIdentifierPartitionManager to make
the tests pass.